add wasm to image; rename artifact mod
